requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
msgpack==1.1.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
pillow==11.0.0
redis==5.2.0
sqlparse==0.5.2
